requirements.txt
numpy
opencv-python
matplotlib
requests
gigachain
gigachat
langchainhub
-U duckduckgo-search
pyfiglet
sympy
climage 
bs4